Added Controller to Ability class Added AllCards static class that does some Reflection voodoo to get all the cards from MtgEngine.*.dll libraries in the engine directory. This allows cards to be added in sets in their own DLLs to avoid ballooning the MtgEngine project with hundreds of classes that aren't necessary to the core functionality of the game. Changed the constructors of abstract classes to protected. Changed the setter for Controller on the Card class to private and added a method to give control to another player. Added Deck class that is capable of parsing deck lists and finding the requisite card classes in the AllCards static class. Added a MtgCardAttribute class to store static attributes on the classes such as Name and Set Name. This will make DeckList parsing easier.
Added AggregateCost for when a card or activated ability has multiple types of costs.

Added TapCost for triggered abilities that require the source be tapped as a cost to activate the ability. This includes mana abilities.

Added TapTargetCost for triggered abilities that require that the controller tap another permanent as a cost to activate an ability.

Added Generic to ManaColor, and added comments explaining some of the nuances with the colors, especially colorless and generic.

Added ManaAmount, ManaParser, and a basic ManaPool. ManaParser still needs to be implemented.

Added ActionBase and PassPriorityAction as actions that Players can take once they have priority. ActionBase is the base from which all Actions must inherit. PassPriorityAction is the action the player chooses when they want to pass priority without performing any further actions.

Added deckList to the Player constructor,  as well as methods to Give a player Priority, and to request that the player Select a Target.

SelectTarget wll likely be further generalized to offer a choice of a set of options, this could also replace GivePriority.

Added ListExtensions class to provide a way to reorder a list to start with a specific item. This is used in the ApNapLoop logic.

Added ReflectionUtils class to get the Constructor for the Card classes defined in the Expansion libraries.

Added a _nextPlayer calculated property to the Game class, as well as an ExtraTurns queue. This may need to be a Stack, I'll worry about that when I get to it.

Added a TestSet expansion library to test the loading of cards from Expansion libraries. Added a Forest card to this that inherits from the BasicLandCard class and implements it for the Green mana color.

Added a PlayerDeck.txt file to the MtgEngineTest project, but it's not currently in use as the deck list is currently hardcoded in the Program.cs file.