Added ActivatedAbilities and ManaAbilities
Added an Add method to the ManaPool class

Combined PlayLand and CastSpell into a single Action type

Added ActivateAbilityAction and PlayCardAction as possible Actions

Fixed a typo in the GivePriority method on the ConsolePlayer class.

Added properties to track number of lands played this turn to the Player class.

Fixed a bug in the NextPlayer calculated property.

Added code to reset the LandsPlayed properties of all players at the beginning of each turn.

Added code to change the active player at the end of each turn.

Added comments and incomplete conditions to the Combat Phase.

Added logic to handle the players chosen actions when they're PlayCardAction or ActivateAbilityAction. This is still unfinished.

Added a basic CheckStateBasedActions method with some comments of things that will need to be checked in this method.