Moved Name, Image, and CardId properties that were on the card and set via the constructor into the MtgCard attribute, and added proxy properties to get those values from the attribute. This will make it easier to keep read-only attributes separate from functional aspects of the card.
Created Card classes for the 5 basic lands and their snow covered variants in the MtgEngine project, and added Inheriting cards to the TestSet library. I have verified that the MtgCard attribute on the inheritng class replaces the MtgCard attribute on the base class, so a new set library can inherit from cards that are reprints.