Reimplemented the ManaPool as a class that encapsulates rather than inherits from a Dictionary. This allows me to be more restrictive about how it can be used to store and access mana as well as allowing me to clear it easily.
Added ManaAbilities to BasicLandCard.

Added Text to the Ability class to make it easier to display to the user their choices in an easy to read.

Added ToString overloads to the Cost classes to make it easier to display the full text string to the user in the ActivatedAbility class.