Added a DrainManaPools method to the Game class that clears out each player's mana pool. This is called at the end of each step (Untap, Upkeep, Draw, Precombat Main, Beginning of Combat, Declare Attackers, Declare Blockers, Damage, End of Combat, Postcombat Main, End Step, and Cleanup).
Fixed a bug where an ApNap loop was being started after activating a mana ability.

Put the Ability class in the Abilities namespace where it should have been already.

Changed the TapTargetCost to have an ActivatedAbility as its source. For now at least, only activated abilities can have costs.

Changed the ActivateAbilityAction to have an ActivatedAbility instead of an Ability. Only activated abilities can be activated.

Fixed some formatting in the text printed in the OfferMulligan method of the ConsolePlayer.