Added a virtual CanCast method to the Card class that, by default, returns true. This method is overloaded by any inheriting class that adds additional stipulations to whether or not a spell can be cast. This includes but is not limited to the requirement that legal targets are available.
Added a Convert method to the ManaCost class, and to the ManaAmount class, and all of these together calculate the CMC for the given mana cost using the values that are stored in the ManaAttribute tag on each mana type.

Changed the OnResolve method of the Effect class to abstract from virtual, because if it does nothing on resolve then it serves no purpose.

Changed the GivePriority method of the Player class to take the Game rather than just the active player so that the player can check the stack to determine possible actions.

Changed the ActivePlayer property of the Game class from private to public with a private setter. This is to allow the player to get the ActivePlayer from the Game class.

Added a method to get the state of the stack in a way that doesn't allow it to be modified outside of the Game class.

Added a CounterSpell method that allows the game to remove a card from the stack.

Added Counterspell to the test set

Added {U}{U} to the hardcoded Mana Parsing code, for now.

Added Islands and Counterspells to the PlayerDeck list.

Added a ChooseTarget method to the player class that enables the player to choose targets for a spell or ability.

Added UnitTests for the ConsolePlayer class to test the ParseChoices method