Moved StaticAbility into the Enums namespace
Implemented Vigilance, FirstStrike, DoubleStrike, Flying, Reach, Haste, and Defender. These still need to be tested. Example creatures need to be added for Vigilance, FirstStrike, DoubleStrike, Haste, and Defender.

Added Rustwing Falcon as an example of a creature with Flying
Added Giang Spider as an example of a creature with Reach.

Made the CanAttack property easier to read, and added IsTapped to the conditions.