Created the MtgEngine.Alpha project for all the cards in the Limited Edition Alpha set.
Added the basic lands, and the dual lands to Alpha.

Changed IAttackable to IDamageable.
Added TakeDamage, and IsDead to IDamageable.
Added LoseLife to Player, which is called by TakeDamage.
Changed Counter class into CounterType enum. It makes counters simpler.

Changed the Blocker selection logic in ConsolePlayer.cs to take into account whether a given creature CAN block the attacker.

Added Power and Toughness properties to the PermanentCard to calculate the total power and total toughness from the BasePower and BaseToughness and the things that modify the power and toughness of the creature.

Added Planeswalkers proxy property to the Zone class.

Added more hardcoded mana parsing conditions to the ManaCost class.