Added the Limited Edition Alpha set.
Added a bunch of cards from Limited Edition Alpha. Not all of them are completely implemented, as I don't have triggered abilities or enchantments implemented.

Added SacrificeSourceCost for ablities that required you to sacrifice the source permanent when you play them. Added SacrificeTargetCost for abilities that require you to sacrifice another permanent to activate them. Added SacrificeCreatureCost and SacrificeArtifactCost which inherit from SacrificeTargetCost and just pass in a premade target selector function.

Removed the ability to play mana abilities while casting a spell because it didn't work with the separation of concerns that I'm going for.

Split some of the bigger classes into multiple partial classes to make the code easier to navigate.