Changed the way player events fire so that the player is the one firing the events and the game subscribes and passes them along as necessary.
Added TookDamage event to the IDamageable interface so that players and creatures both have this event.

Changed the way the ConsolePlayer draws the battlefield to indicate when permanents are tapped.

Fixed some of the logic in the CanAttack property calculation.

Added some logic in the TapCost CanPay method to handle creatures and summoning sickness.

Changed it so that ALL permanents get summoning sickness as they enter the battlefield, but only Creatures are prevented from using tap abilities as a result. If a permanent becomes a creature after entering the battlefield but while it still has summoning sickness then it can no longer use its tap abilities until it either stops being a creature or stops having summoning sickness.