Added Thespian's Stage
Added a Copy abstract method to the Ability class that must be implemented by each non-abstract inheriting class

Added +2/+0 and +0/+2 counter types.

Added IModifier interface and ModifierMode enum.

Added ITargeting interface for abilities that target cards

Added Copy method to all Abilities for when a card copies the abilities of another card.

Added Copy method to all Costs for when a card copies the cost of another card.

Added the PlayerAttemptedToDrawIntoEmptyLibrary flag to the Player class. It gets set when the event occurs and causes the player to lose the game the next time state based actions are checked.