Added the ability to make auras. Currently only EnchantPermanent auras are implemented, and I need to add logic to make them disenchant the cards that they are enchanting when they leave the battlefield.
Added ColorIdentity calculation based on ManaCost. It's in the ManaCost class but it's referenced in the ColorIdentity property getter in the Card class when determining the color identity after the check for devoid.

Added darksteel mutation and Imprisoned in the Moon, but these remain untested.

Updated Marit Lage to add a Modifier that makes it Black. There's no way to hardcode a card's color identity, so it will have to be done with modifiers added at the time of creation.